Fixes comment out assertions
Fixes the assertions about the number of start/terminal vertices. The fix
itself is made at SR.canBeStart(): it removes the condition to have
suitable links as a vertex can be start without having any suitable links
and this is exactly the test assertions (vertices that are both start and
terminal).

Commit: Fix, Mod

TaskNumber: #78, #79